Move seed methods to test host
